% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resegment.R
\name{.classify_cna_segments}
\alias{.classify_cna_segments}
\title{Internal: Classify CNA Segments}
\usage{
.classify_cna_segments(
  filt,
  l3,
  l4,
  low_gain,
  normal_gain,
  high_gain,
  low_loss,
  normal_loss,
  high_loss,
  chrom_percent,
  arm_percent,
  focal_percent_low,
  focal_percent_medium,
  focal_percent_high,
  min_baf
)
}
\description{
Internal: Classify CNA Segments
}
\keyword{internal}
