% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resegment.R
\name{.classify_intensity}
\alias{.classify_intensity}
\title{Classify CNA by Intensity}
\usage{
.classify_intensity(w)
}
\arguments{
\item{w}{Numeric intensity (1=low, 2=medium, 3=high).}
}
\value{
Character intensity label.
}
\description{
Converts numeric intensity code (1-3) to string label.
}
\keyword{internal}
