% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resegment.R
\name{.resegment_iterative}
\alias{.resegment_iterative}
\title{Internal: Iterative Re-segmentation}
\usage{
.resegment_iterative(
  file,
  min_length,
  max_dist_segm,
  percent_dist,
  dev_btw_segs,
  dev_tozero,
  dev_baf
)
}
\description{
Internal: Iterative Re-segmentation
}
\keyword{internal}
