% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{load_prediction_model}
\alias{load_prediction_model}
\title{Load Prediction Model}
\usage{
load_prediction_model(model_name)
}
\arguments{
\item{model_name}{Character model file name (without .RData extension).
Should be located in inst/models/ directory.}
}
\value{
A random forest model object.
}
\description{
Loads a pre-trained random forest model for CNA-based classification.
}
\details{
Models should be trained and saved as .RData files containing a
variable named "model_cms" or equivalent.
}
\keyword{internal}
