% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resegment.R
\name{classify_cna}
\alias{classify_cna}
\title{Classify Single CNA Based on seg.mean}
\usage{
classify_cna(
  seg_mean,
  low_gain,
  medium_gain,
  high_gain,
  low_loss,
  medium_loss,
  high_loss
)
}
\arguments{
\item{seg_mean}{Numeric log2 ratio.}

\item{low_gain}{Numeric low gain cutoff.}

\item{medium_gain}{Numeric medium gain cutoff.}

\item{high_gain}{Numeric high gain cutoff.}

\item{low_loss}{Numeric low loss cutoff.}

\item{medium_loss}{Numeric medium loss cutoff.}

\item{high_loss}{Numeric high loss cutoff.}
}
\value{
Integer (0=no alteration, 1=low intensity, 2=medium, 3=high).
}
\description{
Internal function that classifies a segment as gain/loss and returns intensity.
}
